add userId check to MealRepository

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -8,19 +8,23 @@ import java.util.Collection;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
 public class InMemoryMealRepositoryImpl implements MealRepository {
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
-    private AtomicInteger counter = new AtomicInteger(0);
+    private Integer counter = 0;
 
     {
-        MealsUtil.MEALS.forEach(this::save);
+        MealsUtil.MEALS.forEach(meal -> save(meal.getUserId(),meal));
     }
 
     @Override
-    public Meal save(Meal meal) {
+    public Meal save(int userId, Meal meal) {
+        if (userId != meal.getUserId()){
+            return null;
+        }
         if (meal.isNew()) {
-            meal.setId(counter.incrementAndGet());
+            meal.setId(counter++);
             repository.put(meal.getId(), meal);
             return meal;
         }
@@ -29,18 +33,27 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
     }
 
     @Override
-    public void delete(int id) {
+    public void delete(int userId, int id) {
+        if (get(userId, id).getUserId() != userId){
+            return;
+        }
         repository.remove(id);
     }
 
     @Override
-    public Meal get(int id) {
-        return repository.get(id);
+    public Meal get(int userId, int id) {
+        Meal meal = repository.get(id);
+        return userId == meal.getUserId() ? meal : null;
     }
 
     @Override
-    public Collection<Meal> getAll() {
-        return repository.values();
+    public Collection<Meal> getAll(int userId) {
+        return repository
+                .values()
+                .stream()
+                .filter(meal -> meal.getUserId()==userId)
+                .sorted((m1,m2) -> m2.getDateTime().compareTo(m1.getDateTime()))
+                .collect(Collectors.toList());
     }
 }
